Use `shard_head_state` to link tx to events
In `shard_head_state`, we use log_listeners to listen to the events.
If a tx emit logs(an event occurs), we can use `self.last_tx` to get the
`tx_hash` of the event.

In mainchain we don't need to do this. Because we only trust the longest
chain in mainchain. We don't care about forks in mainchain.